bugs fixed

+++ route/newEmployee.js
@@ -9,59 +9,67 @@ router.post("/", async (req, res) => {
    // console.log(req.body);
     const user1 = new User1();  // creating an object of an class User1
     user1.employeename = req.body.employeename;
-    user1.dummy=false
-    await user1.save();
+    user1.dummy = req.body.dummy;
+
+  await user1.save();
     res.send(user1);
 });
 
 router.get("/", async (req, res) => {
     const user1 = await User1.find({})
-    res.send(user1)
+  console.log(user1)
+  let employee =[]
+  user1.forEach(item=>{
+    if(item.dummy===false){
+      employee.push(item)
+    }
+  })
+    res.send(employee)
 });
 // Delete Operatio
 
 
 router.delete("/:user1Id", async (req, res) => {
   const user = await User1.findById({ _id: req.params.user1Id}).exec();
-   if(user.dummy==false){
+  console.log(user)
+
   const assign = await Assign.find({ employeename: req.params.user1Id }).exec();
-  let taskId=[]
-  assign.forEach((item)=>{
-    taskId.push(item.task)
-  })
-  for(i=0;i<taskId.length;i++){
-    await Task.findByIdAndRemove({_id: taskId[i]});
-  }
-  for(i=0;i<assign.length;i++){
-    await Assign.findByIdAndRemove({_id: assign[i]._id});
-  }
+
+     let taskId=[]
+     if(assign.length>0) {
+       assign.forEach((item) => {
+         taskId.push(item.task)
+       })
+       for (i = 0; i < taskId.length; i++) {
+         await Task.findByIdAndRemove({_id: taskId[i]});
+       }
+       for (i = 0; i < assign.length; i++) {
+         await Assign.findByIdAndRemove({_id: assign[i]._id});
+       }
+     }
+
   const user1 = await Emp.find({ employeeName: req.params.user1Id}).exec();
-  let empId=[]
-  user1.forEach((item)=>{
-    empId.push(item._id)
-  })
-  obj={"employeeName":'5f3f59d5d832a537cc8f7529'}
- // console.log(empId[0])
-  const use = await Emp.findByIdAndUpdate({
-    _id: empId[0]
-},
+     if(user1.length>0){
+       let empId=[]
+       user1.forEach((item)=>{
+         empId.push(item._id)
+       })
+       obj={"employeeName":'5f49c19d5ad27849a06df32b'}
+       const use = await Emp.findByIdAndUpdate({
+           _id: empId[0]
+         }, obj, {
+           new: true,
+           runValidators: true
+         });
+     }
 
-
-obj, {
-    new: true,
-    runValidators: true
-});
   //console.log(empId)
-  
-  const user2 = await Emp.findByIdAndRemove({
+
+  const user2 = await User1.findByIdAndRemove({
       _id: req.params.user1Id
     });
-  //console.log("ABCDEFGH")
-    res.send(user2)
-}
-// else
-// res.send(user.dummy)
-// console.log("EFGHIJKL"+user.dummy)
+    res.send({result:"data deleated"})
+
  })
 
 //GetElementById
